Update CategorySummary in preparation for CategoryDetailView
