+++ budget-binder-multiplatform-app/src/jsMain/kotlin/de/hsfl/budgetBinder/compose/Composables.kt
@@ -180,95 +180,7 @@ fun CategoryImagesToImageList(
     }
 }
 
-@OptIn(ExperimentalComposeWebSvgApi::class)
-@Composable
-fun CategoryList(
-    categoryList: List<Category>,
-    onEditButton: (Int) -> Unit,
-    onDeleteButton: (Int) -> Unit
-) {
-    var deleteDialog by remember { mutableStateOf(false) }
-    var id by remember { mutableStateOf(0) }
 
-    Div {
-        if (deleteDialog) {
-            DeleteDialog(
-                false,
-                { onDeleteButton(id) },
-                { deleteDialog = false }) { Text("Delete Category?") }
-        }
-        for (category in categoryList)
-            Div(attrs = {
-                classes("mdc-card", AppStylesheet.card)
-            }
-            ) {
-                Div(
-                    attrs = {
-                        classes(AppStylesheet.categoryListElement, AppStylesheet.flexContainer)
-                    }
-                ) {
-                    Div(
-                        attrs = {
-                            classes(AppStylesheet.imageFlexContainer)
-                        }
-                    ) {
-                        CategoryImageToIcon(category.image)
-                    }
-                    Div(
-                        attrs = {
-                            classes(AppStylesheet.categoryListElementText)
-                        }
-                    ) {
-                        Div {
-                            Div(attrs = {
-                                classes("mdc-typography--headline4", AppStylesheet.text)
-                            }) { Text(category.name) }
-                            Div(attrs = {
-                                classes("mdc-typography--headline6", AppStylesheet.text)
-                            }) { Text("Budget: ${category.budget}€") }
-                        }
-                    }
-                    Div(attrs = {
-                        classes(AppStylesheet.imageFlexContainer)
-                    }
-                    ) {
-                        Svg(viewBox = "0 0 1 1") {//For aspect ratio - tries to fill out wherever it is in
-                            Circle(cx = 0.5, cy = 0.5, r = 0.5, {
-                                attr("fill", "#${category.color}")
-                            })
-                        }
-                    }
-                }
-                Div(
-                    attrs = {
-                        classes(AppStylesheet.flexContainer)
-                    }
-                ) {
-                    Button(attrs = {
-                        classes("mdc-button", "mdc-button--raised", AppStylesheet.marginRight)
-                        onClick { onEditButton(category.id) }
-                        style {
-                            flex(50.percent)
-                            margin(1.5.percent)
-                        }
-                    }) {
-                        Text("Edit Category")
-                    }
-                    Button(attrs = {
-                        classes("mdc-button", "mdc-button--raised")
-                        onClick { deleteDialog = !deleteDialog; id = category.id }
-                        style {
-                            flex(50.percent)
-                            margin(1.5.percent)
-                            backgroundColor(Color("#b00020"))
-                        }
-                    }) {
-                        Text("Delete Category")
-                    }
-                }
-            }
-    }
-}
 
 @Composable
 fun DeleteDialog(
